requirements.txt
Django==1.8.2
django-crispy-forms==1.4.0
django-registration-redux==1.2
mysqlclient==1.3.6
Pillow==2.9.0
selenium==2.46.0
sorl-thumbnail==12.3
wheel==0.24.0
